Refactored into LocationService.php
